requirements.txt
aiohttp==3.8.4
aiosignal==1.3.1
asgiref==3.6.0
async-timeout==4.0.2
attrs==23.1.0
backports.zoneinfo==0.2.1
certifi==2022.12.7
charset-normalizer==3.1.0
colorama==0.4.6
dataclasses-json==0.5.7
Django==4.2
django-environ==0.10.0
dnspython==2.3.0
frozenlist==1.3.3
greenlet==2.0.2
idna==3.4
langchain==0.0.161
loguru==0.7.0
marshmallow==3.19.0
marshmallow-enum==1.5.1
multidict==6.0.4
mypy-extensions==1.0.0
mysqlclient==2.1.1
numexpr==2.8.4
numpy==1.22.4
openai==0.27.6
openapi-schema-pydantic==1.2.4
packaging==23.1
pinecone-client==2.2.1
pydantic==1.10.7
python-dateutil==2.8.2
PyYAML==6.0
requests==2.29.0
six==1.16.0
SQLAlchemy==2.0.10
sqlparse==0.4.4
tenacity==8.2.2
tqdm==4.65.0
typing-extensions==4.5.0
typing-inspect==0.8.0
tzdata==2023.3
urllib3==1.26.15
whitenoise==6.4.0
win32-setctime==1.1.0
yarl==1.9.2
